pyproject.toml
[project]
name = "percona-dk"
version = "0.1.0"
description = "Percona Developer Knowledge — canonical, structured access to Percona documentation for AI assistants and tools"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "chromadb>=1.0",
    "fastmcp>=2.0",
    "gitpython>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
